refactor: improve main.py structure and add type safety
- Refactored main.py from 556 to ~205 lines by extracting code into focused modules
- Created data_collection.py for data collection functions and context classes
- Created metrics_calculation.py for metrics calculation logic
- Created output_formatting.py for results table generation and formatting
- Added CommandLineArgs TypedDict for structured command line argument handling
- Improved type safety with proper type annotations throughout
- Maintained all existing functionality - all 59 tests pass
- Simplified function signatures by using structured arguments instead of 10+ parameters
